chore: refactor so data doesn't go down, up, down, down
